+++ Enrollment System 2.0/AdminSubjectPage.cs
@@ -90,22 +90,30 @@ namespace Enrollment_System_2._0
         {
             try
             {
-                id = int.Parse(dataGridView1.CurrentRow.Cells[1].Value.ToString());
-                subcode.Text = dataGridView1.CurrentRow.Cells[2].Value.ToString();
-                subname.Text = dataGridView1.CurrentRow.Cells[3].Value.ToString();
-                subunits.Text = dataGridView1.CurrentRow.Cells[4].Value.ToString();
-                subyear.Text = dataGridView1.CurrentRow.Cells[5].Value.ToString();
-                comboBox1.Text = dataGridView1.CurrentRow.Cells[6].Value.ToString();
-                if (dataGridView1.Columns[e.ColumnIndex].Name == "remove")
+                if (dataGridView1.CurrentRow.Index == dataGridView1.Rows.Count - 1)
                 {
-                    if (MessageBox.Show("Confirm?", "Message", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+                    MessageBox.Show("No records found!", "Message");
+                }
+                else
+                {
+                    id = int.Parse(dataGridView1.CurrentRow.Cells[1].Value.ToString());
+                    subcode.Text = dataGridView1.CurrentRow.Cells[2].Value.ToString();
+                    subname.Text = dataGridView1.CurrentRow.Cells[3].Value.ToString();
+                    subunits.Text = dataGridView1.CurrentRow.Cells[4].Value.ToString();
+                    subyear.Text = dataGridView1.CurrentRow.Cells[5].Value.ToString();
+                    comboBox1.Text = dataGridView1.CurrentRow.Cells[6].Value.ToString();
+                    if (dataGridView1.Columns[e.ColumnIndex].Name == "remove")
                     {
-                        db.delete_subject(id);
-                        dataGridView1.DataSource = db.view_subject();
-                        MessageBox.Show("Subject deleted sucessfully", "Message");
-                        ClearData();
+                        if (MessageBox.Show("Confirm?", "Message", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+                        {
+                            db.delete_subject(id);
+                            dataGridView1.DataSource = db.view_subject();
+                            MessageBox.Show("Subject deleted sucessfully", "Message");
+                            ClearData();
+                        }
                     }
                 }
+               
             }
             catch (Exception)
             {
